made MultiKeyDicts (sort of) "chainable" [FLAKY]
As noted above, the changes introduced by this commit are very flaky (i.e. a rollback is more probable than usual).

The basic problem is that the "maxdepth" attribute of the MultiKeyDict was serving two functions that should really be distinct: 1) determining which factory (leaffactory or the default "branchfactory") to invoke when a key is not found; and 2) enforcing the maximum length of the tuples that can be passed to the get and set methods.  This means that a MKD that happens to have other MKD objects at its 'nominal' leaves will bomb if its get method receives as argument a tuple that includes key elements for both the 'enclosing' and the 'enclosed' MKDs...  E.g.

  submkd = MultiKeyDict(maxdepth=2)
  submkd.set(('c', 'd')) = 42
  root = MultiKeyDict(maxdepth=2)
  root.set(('a', 'b')) = submkd

  # I would like to be able to now do something like
  print root.get(('a', 'b', 'c', 'd'))

  # but, as things stood before, the last line would have triggered a
  KeyError: "('a', 'b', 'c', 'd') (exceeds maxdepth=2)"

In fact, the following also would have produced an error, for the same reason

  root = MultiKeyDict(maxdepth=2)
  root.set(('a', 'b')) = {'c': 21}
  print root.get(('a', 'b', 'c'))

  --> KeyError: "('a', 'b', 'c') (exceeds maxdepth=2)"

The idea is to make it possible to use a (subclass of) MKD, or simply a (subclass of) dict, as the "leaf class" for an enclosing MKD, in such a way that the resulting MKD has, effectively, a greater depth than what's specified by its 'nominal' "maxdepth" attribute.  This nominal "maxdepth" attribute is still necessary for the "enclosing" MKD to know when it must invoke the leaffactory (upon failure to find a key), but it no longer determines the maximum length of the tuples that can be passed to the get and set methods.  This is now done by the new "height" attribute, which depends on the value of the "maxdepth" attribute and of the "height" attribute of the object returned by the leaffactory.  (If the latter is None, then the "height" attribute of the enclosing MKD is also None.)

It follows from the above that the name "maxdepth" is actually a misnomer; I will change it soon (I will probably change "height" also).

Then again, the implementation of this enhanced capability is VERY FLAKY, so the whole thing may need to be rolled back pretty soon...