"fixed" bug in implementation of Hdf5File
This fix entailed implementing a fileno method for the Hdf5File class.  The implementation I came up with keeps flock from throwing an exception, but this implementation was just a wild guess, since I was not able to find any documentation for the attribute it uses.  Hdf5File SHOULD BE REGARDED BROKEN UNTIL FURTHER NOTICE.  In particular, concurrent access by multiple processes to a single file may result in data corruption, and/or filesystem problems.