script to save pickled cubes to hdf5 file
(The name of the script, icbp45_mergecubes.py, is an anachronism.  It should be called something like icbp45_pkl2h5.py.)

The script reads in pickled data and confounder cubes, massages the former if necessary (to achieve adequate cubosity), and stuffs them into an HDF5 file.  At the moment, the confounder data cube is saved as a (massive) pickled MultiKeyDict object, although the plan is to change this in the near-future.