change some barcode cols to strings
Changed the contents of the "barcode" column of the sheets PlateData
and WellDataMapped, so that the couple of erroneous barcodes, which
appear as date-time cells, will now contain plain strings, in the
format '%Y%m%dT%H%M%S'.

(This was done by using the formula =TEXT(<CELL>, "YYYYMMDDThhmmss").)

The reason for doing this is that the proper handling of these faulty
barcodes in MATLAB is too difficult to be worth the trouble.  (The
problem is that MATLAB reads these barcode columns as strings, which
results in a loss of precision for the datetime cells; as a result,
converting these values to date strings results in dates that are a
couple of seconds off.  This then becomes a problem when attempting to
match these "barcodes" with the corresponding barcodes from other
spreadsheets, which are *originally* stored as strings.)